# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/san-pham/danh-muc/{danhMucId}": {
      "get": {
        "description": "Trả về danh sách sản phẩm thuộc một danh mục cụ thể",
        "operationId": "SanPhamController_findByDanhMuc",
        "parameters": [
          {
            "name": "danhMucId",
            "required": true,
            "in": "path",
            "description": "ID danh mục",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy sản phẩm theo danh mục thành công"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy sản phẩm theo danh mục",
        "tags": [
          "Sản phẩm"
        ]
      }
    },
    "/san-pham/{id}": {
      "get": {
        "description": "Trả về thông tin chi tiết của một sản phẩm bao gồm hình ảnh và đánh giá",
        "operationId": "SanPhamController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID sản phẩm",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin sản phẩm thành công"
          },
          "404": {
            "description": "Không tìm thấy sản phẩm"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin chi tiết sản phẩm",
        "tags": [
          "Sản phẩm"
        ]
      },
      "delete": {
        "description": "Xóa sản phẩm khỏi hệ thống",
        "operationId": "SanPhamController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID sản phẩm cần xóa",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa sản phẩm thành công"
          },
          "404": {
            "description": "Không tìm thấy sản phẩm"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa sản phẩm",
        "tags": [
          "Sản phẩm"
        ]
      }
    },
...
    "/danh-muc": {
      "post": {
        "operationId": "DanhMucController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDanhMucDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "DanhMuc"
        ]
      },
      "get": {
        "operationId": "DanhMucController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DanhMuc"
        ]
      }
    },
    "/danh-muc/{tendanhmuc}": {
      "get": {
        "operationId": "DanhMucController_findOne",
        "parameters": [
          {
            "name": "tendanhmuc",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DanhMuc"
        ]
      }
    },
    "/danh-muc/{id}": {
      "patch": {
        "operationId": "DanhMucController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDanhMucDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DanhMuc"
        ]
      },
      "delete": {
        "operationId": "DanhMucController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DanhMuc"
        ]
      }
    },
...
    "/danh-gia": {
      "post": {
        "operationId": "DanhGiaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDanhGiaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đánh giá được tạo thành công"
          },
          "400": {
            "description": "Dữ liệu đầu vào không hợp lệ hoặc đã đánh giá sản phẩm"
          },
          "403": {
            "description": "Không có quyền hoặc chưa mua sản phẩm này"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo đánh giá mới (Chỉ khách hàng)",
        "tags": [
          "Đánh giá"
        ]
      }
    },
    "/danh-gia/san-pham/{sanPhamId}": {
      "get": {
        "operationId": "DanhGiaController_getBySanPham",
        "parameters": [
          {
            "name": "sanPhamId",
            "required": true,
            "in": "path",
            "description": "ID sản phẩm",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách thành công"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy đánh giá theo ID sản phẩm",
        "tags": [
          "Đánh giá"
        ]
      }
    },
    "/danh-gia/khach-hang/{khachHangId}": {
      "get": {
        "operationId": "DanhGiaController_getByKhachHang",
        "parameters": [
          {
            "name": "khachHangId",
            "required": true,
            "in": "path",
            "description": "ID khách hàng",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách thành công"
          },
          "403": {
            "description": "Không có quyền truy cập"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy đánh giá theo ID khách hàng",
        "tags": [
          "Đánh giá"
        ]
      }
    },
    "/danh-gia/rating/{sanPhamId}": {
      "get": {
        "operationId": "DanhGiaController_getSanPhamRating",
        "parameters": [
          {
            "name": "sanPhamId",
            "required": true,
            "in": "path",
            "description": "ID sản phẩm",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin đánh giá thành công"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin đánh giá trung bình của sản phẩm",
        "tags": [
          "Đánh giá"
        ]
      }
    },
    "/danh-gia/{id}": {
      "get": {
        "operationId": "DanhGiaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID đánh giá",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin thành công"
          },
          "404": {
            "description": "Không tìm thấy đánh giá"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin chi tiết đánh giá theo ID",
        "tags": [
          "Đánh giá"
        ]
      },
      "delete": {
        "operationId": "DanhGiaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID đánh giá cần xóa",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa thành công"
          },
          "403": {
            "description": "Không có quyền xóa"
          },
          "404": {
            "description": "Không tìm thấy đánh giá"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa đánh giá",
        "tags": [
          "Đánh giá"
        ]
      }
    },
...
    "/khach-hang/{id}": {
      "get": {
        "operationId": "KhachHangController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KhachHang"
        ]
      },
      "delete": {
        "operationId": "KhachHangController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KhachHang"
        ]
      }
    },
    "/don-hang/cart/add": {
      "post": {
        "operationId": "DonHangController_addToCart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Thêm vào giỏ hàng thành công"
          },
          "400": {
            "description": "Sản phẩm không tồn tại hoặc không đủ số lượng"
          },
          "403": {
            "description": "Chỉ khách hàng mới có quyền thêm vào giỏ hàng"
          }
        },
        "summary": "Thêm sản phẩm vào giỏ hàng (Khách hàng)",
        "tags": [
          "DonHang"
        ]
      }
    },
    "/don-hang/cart": {
      "get": {
        "operationId": "DonHangController_getCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lấy giỏ hàng thành công"
          },
          "403": {
            "description": "Chỉ khách hàng mới có quyền xem giỏ hàng"
          }
        },
        "summary": "Xem giỏ hàng (Khách hàng)",
        "tags": [
          "DonHang"
        ]
      }
    },
    "/don-hang/cart/update/{maSanPham}/{soLuong}": {
      "patch": {
        "operationId": "DonHangController_updateCartItem",
        "parameters": [
          {
            "name": "maSanPham",
            "required": true,
            "in": "path",
            "description": "ID sản phẩm",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "soLuong",
            "required": true,
            "in": "path",
            "description": "Số lượng mới",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cập nhật giỏ hàng thành công"
          },
          "400": {
            "description": "Không đủ số lượng trong kho"
          },
          "404": {
            "description": "Sản phẩm không có trong giỏ hàng"
          }
        },
        "summary": "Cập nhật số lượng sản phẩm trong giỏ hàng (Khách hàng)",
        "tags": [
          "DonHang"
        ]
      }
    },
    "/don-hang/cart/remove/{maSanPham}": {
      "delete": {
        "operationId": "DonHangController_removeFromCart",
        "parameters": [
          {
            "name": "maSanPham",
            "required": true,
            "in": "path",
            "description": "ID sản phẩm cần xóa",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa sản phẩm khỏi giỏ hàng thành công"
          },
          "404": {
            "description": "Sản phẩm không có trong giỏ hàng"
          }
        },
        "summary": "Xóa sản phẩm khỏi giỏ hàng (Khách hàng)",
        "tags": [
          "DonHang"
        ]
      }
    },
...
    "/don-hang/cart/checkout": {
      "post": {
        "operationId": "DonHangController_checkoutCart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Thanh toán thành công"
          },
          "400": {
            "description": "Giỏ hàng trống hoặc sản phẩm không đủ số lượng"
          },
          "404": {
            "description": "Không tìm thấy giỏ hàng"
          }
        },
        "summary": "Thanh toán giỏ hàng (Khách hàng)",
        "tags": [
          "DonHang"
        ]
      }
    }
  },
  "info": {
    "title": "API Quản lý Siêu Thị",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDanhMucDto": {
        "type": "object",
        "properties": {
          "tenDanhMuc": {
            "type": "string",
            "description": "Tên danh mục"
          },
          "moTa": {
            "type": "string",
            "description": "Mô tả danh mục"
          }
        }
      },
      "UpdateDanhMucDto": {
        "type": "object",
        "properties": {
          "tenDanhMuc": {
            "type": "string",
            "description": "Tên danh mục"
          },
          "moTa": {
            "type": "string",
            "description": "Mô tả danh mục"
          }
        }
      },
...
      "CreateDanhGiaDto": {
        "type": "object",
        "properties": {
          "soSao": {
            "type": "number",
            "description": "Số sao đánh giá (1-5)",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "noiDung": {
            "type": "string",
            "description": "Nội dung đánh giá",
            "example": "Sản phẩm rất tốt, giao hàng nhanh chóng!"
          },
          "sanPhamId": {
            "type": "number",
            "description": "ID sản phẩm được đánh giá",
            "example": 1
          }
        },
        "required": [
          "soSao",
          "noiDung",
          "sanPhamId"
        ]
      },
...
      "AddToCartDto": {
        "type": "object",
        "properties": {
          "sanPhamId": {
            "type": "number",
            "description": "ID sản phẩm",
            "example": 1
          },
          "soLuong": {
            "type": "number",
            "description": "Số lượng",
            "example": 2,
            "minimum": 1
          }
        },
        "required": [
          "sanPhamId",
          "soLuong"
        ]
      },
      "CheckoutCartDto": {
        "type": "object",
        "properties": {
          "tenDonHang": {
            "type": "string"
          },
          "ngayChuanBiHang": {
            "format": "date-time",
            "type": "string"
          },
          "ngayNhanHang": {
            "format": "date-time",
            "type": "string"
          },
          "donViVanChuyen": {
            "type": "string"
          },
          "phuongThucThanhToan": {
            "type": "string"
          },
          "trangThaiDonHang": {
            "type": "string"
          },
          "diaChiGiaoHang": {
            "type": "string",
            "description": "Địa chỉ giao hàng"
          },
          "sdtNguoiNhan": {
            "type": "string",
            "description": "Số điện thoại người nhận"
          },
          "tenNguoiNhan": {
            "type": "string",
            "description": "Tên người nhận"
          },
          "ghiChu": {
            "type": "string",
            "description": "Ghi chú"
          }
        },
        "required": [
          "tenDonHang",
          "ngayChuanBiHang",
          "ngayNhanHang",
          "donViVanChuyen",
          "phuongThucThanhToan",
          "trangThaiDonHang",
          "diaChiGiaoHang",
          "sdtNguoiNhan",
          "tenNguoiNhan",
          "ghiChu"
        ]
      }
    }
  }
}